fix: Corrigir edição de quantidade e melhorar tratamento de erros
- Mudar step do campo quantidade para 1 (inteiro) ao invés de 0.001
- Usar parseInt ao invés de parseFloat para quantidade
- Adicionar tratamento de valores null/undefined (location, notes, avgCost)
- Melhorar logs de erro com stack trace e mensagem detalhada
- Adicionar log de request body para debug

+++ backend/public/script.js
@@ -583,7 +583,7 @@ class PhotoInventoryManager {
         const category = document.getElementById('editProductCategory').value;
         const name = document.getElementById('editProductName').value.trim();
         const unit = document.getElementById('editProductUnit').value.trim();
-        const quantity = parseFloat(document.getElementById('editProductQuantity').value);
+        const quantity = parseInt(document.getElementById('editProductQuantity').value);
         const minStock = parseInt(document.getElementById('editProductMinStock').value);
         const cost = parseFloat(document.getElementById('editProductCost').value) || 0;
         const location = document.getElementById('editProductLocation').value.trim();
